Move init tests to test_init.py
* Fix default for DEFAULT_PATH
