+++ custom_components/mail_and_packages/const.py
@@ -39,7 +39,7 @@ CONF_AMAZON_FWDS = "amazon_fwds"
 DEFAULT_NAME = "Mail And Packages"
 DEFAULT_PORT = "993"
 DEFAULT_FOLDER = '"INBOX"'
-DEFAULT_PATH = "/images/mail_and_packages/"
+DEFAULT_PATH = "images/mail_and_packages/"
 DEFAULT_IMAGE_SECURITY = True
 DEFAULT_IMAP_TIMEOUT = 30
 DEFAULT_GIF_DURATION = 5
