+++ tests/test_helpers.py
@@ -1,4 +1,4 @@
-"""Tests for init module."""
+"""Tests for helpers module."""
 import datetime
 from datetime import date
 from unittest import mock
@@ -33,85 +33,11 @@ from tests.const import (
     FAKE_CONFIG_DATA_CORRECTED,
     FAKE_CONFIG_DATA_CORRECTED_EXTERNAL,
     FAKE_CONFIG_DATA_EXTERNAL,
-    FAKE_CONFIG_DATA_NO_PATH,
     FAKE_CONFIG_DATA_NO_RND,
 )
 
 MAIL_IMAGE_URL_ENTITY = "sensor.mail_image_url"
 MAIL_IMAGE_SYSTEM_PATH = "sensor.mail_image_system_path"
-
-
-async def test_unload_entry(hass, mock_update, mock_copy_overlays):
-    """Test unloading entities. """
-    entry = MockConfigEntry(
-        domain=DOMAIN,
-        title="imap.test.email",
-        data=FAKE_CONFIG_DATA,
-    )
-
-    entry.add_to_hass(hass)
-    assert await hass.config_entries.async_setup(entry.entry_id)
-    await hass.async_block_till_done()
-
-    assert len(hass.states.async_entity_ids(SENSOR_DOMAIN)) == 28
-    entries = hass.config_entries.async_entries(DOMAIN)
-    assert len(entries) == 1
-
-    assert await hass.config_entries.async_unload(entries[0].entry_id)
-    assert len(hass.states.async_entity_ids(SENSOR_DOMAIN)) == 0
-    assert len(hass.states.async_entity_ids(DOMAIN)) == 0
-
-
-async def test_setup_entry(
-    hass,
-    mock_imap_no_email,
-    mock_osremove,
-    mock_osmakedir,
-    mock_listdir,
-    mock_update_time,
-    mock_copy_overlays,
-    mock_hash_file,
-    mock_getctime_today,
-):
-    """Test settting up entities. """
-    entry = MockConfigEntry(
-        domain=DOMAIN,
-        title="imap.test.email",
-        data=FAKE_CONFIG_DATA,
-    )
-
-    entry.add_to_hass(hass)
-    assert await hass.config_entries.async_setup(entry.entry_id)
-    await hass.async_block_till_done()
-
-    assert len(hass.states.async_entity_ids(SENSOR_DOMAIN)) == 28
-    entries = hass.config_entries.async_entries(DOMAIN)
-    assert len(entries) == 1
-
-
-async def test_no_path_no_sec(
-    hass,
-    mock_imap_no_email,
-    mock_osremove,
-    mock_osmakedir,
-    mock_listdir,
-    mock_update_time,
-    mock_copy_overlays,
-    mock_hash_file,
-    mock_getctime_today,
-):
-    """Test settting up entities. """
-    entry = MockConfigEntry(
-        domain=DOMAIN, title="imap.test.email", data=FAKE_CONFIG_DATA_NO_PATH, version=3
-    )
-
-    entry.add_to_hass(hass)
-    assert await hass.config_entries.async_setup(entry.entry_id)
-    await hass.async_block_till_done()
-
-    assert len(hass.states.async_entity_ids(SENSOR_DOMAIN)) == 28
-    entries = hass.config_entries.async_entries(DOMAIN)
-    assert len(entries) == 1
 
 
 async def test_get_formatted_date():
